requirements.txt
openai==0.28
flask==3.0.0
load_dotenv